# crates/server/src/cypher_tests.rs
//! Production transport boundaries share one Cypher service and database.
use crate::{grpc, http, state};
use axum::{
    body::{to_bytes, Body},
    http::{Request, StatusCode},
};
use grpc::pb::helix_db_server_server::HelixDbServer;
use serde_json::json;
use std::sync::Arc;
use tower::ServiceExt;

#[tokio::test]
async fn cypher_http_and_grpc_preserve_values_errors_and_atomic_writes() {
    let db = Arc::new(
        db::HelixDB::open(db::HelixDbSource::InMemory {
            database: "cypher-transport".into(),
        })
        .await
        .unwrap(),
    );
    let state = state::ServerState::new(Arc::clone(&db), None, crate::StorageConfig::Memory)
        .with_cypher_endpoints(crate::CypherEndpoints::Enabled);
    let router = http::router(state.clone());
    let grpc = grpc::GrpcService::new(state);
    let response = router.clone().oneshot(Request::post("/v2/cypher").header("content-type","application/json")
        .body(Body::from(json!({"query":"CREATE (n:Transport {name:$name}) RETURN n.name AS name, $large AS large", "parameters":{"name":"Ada","large":{"$type":"integer","value":"9223372036854775807"}}}).to_string())).unwrap()).await.unwrap();
    assert_eq!(response.status(), StatusCode::OK);
    let value: serde_json::Value =
        serde_json::from_slice(&to_bytes(response.into_body(), 4096).await.unwrap()).unwrap();
    assert_eq!(
        value,
        json!({"columns":["name","large"],"rows":[["Ada",{"$type":"integer","value":"9223372036854775807"}]]})
    );

    let result = grpc
        .execute_cypher(tonic::Request::new(grpc::pb::QueryJsonRequest {
            body: json!({"query":"MATCH (n:Transport) RETURN n.name AS name"})
                .to_string()
                .into_bytes()
                .into(),
            ..Default::default()
        }))
        .await
        .unwrap()
        .into_inner();
    assert_eq!(
        serde_json::from_slice::<serde_json::Value>(&result.body).unwrap(),
        json!({"columns":["name"],"rows":[["Ada"]]})
    );

    let response = router.clone().oneshot(Request::post("/v2/cypher").body(Body::from(json!({"query":"CREATE (:Transport {name:'rollback'}) WITH 1 AS x RETURN 1 / (x - x)"}).to_string())).unwrap()).await.unwrap();
    assert_eq!(response.status(), StatusCode::BAD_REQUEST);
    let error: serde_json::Value =
        serde_json::from_slice(&to_bytes(response.into_body(), 4096).await.unwrap()).unwrap();
    assert_eq!(error["details"]["phase"], "runtime");
    assert_eq!(error["details"]["detail"], "DivisionByZero");

    let error = grpc
        .execute_cypher(tonic::Request::new(grpc::pb::QueryJsonRequest {
            body: json!({"query":"RETURN missing"})
                .to_string()
                .into_bytes()
                .into(),
            ..Default::default()
        }))
        .await
        .unwrap_err();
    assert_eq!(error.code(), tonic::Code::InvalidArgument);
    let details: helix_planner::relational::QueryError =
        serde_json::from_slice(error.details()).unwrap();
    assert_eq!(details.detail, "UndefinedVariable");
    assert_eq!(
        details.phase,
        helix_planner::relational::ErrorPhase::Compile
    );

    let response = router
        .clone()
        .oneshot(
            Request::post("/v2/cypher")
                .header("x-helix-warm", "true")
                .body(Body::from(
                    json!({"query":"CREATE (:Transport {name:'warm'})"}).to_string(),
                ))
                .unwrap(),
        )
        .await
        .unwrap();
    assert_eq!(response.status(), StatusCode::BAD_REQUEST);
    let result = db
        .cypher(db::cypher::Request::new(
            "MATCH (n:Transport) RETURN count(n)",
        ))
        .await
        .unwrap();
    assert_eq!(result.rows, vec![vec![json!(1)]]);
    drop(router);
    drop(grpc);
    db.close().await.unwrap();
}

#[tokio::test]
async fn cypher_explain_has_a_separate_read_only_http_contract() {
    let source = db::HelixDbSource::InMemory {
        database: "cypher-explain-transport".into(),
    };
    let config = source
        .embedded_default_config()
        .with_query_telemetry(db::config::QueryTelemetry::Disabled);
    let db = Arc::new(db::HelixDB::open_with_config(source, config).await.unwrap());
    let router = http::router(
        state::ServerState::new(Arc::clone(&db), None, crate::StorageConfig::Memory)
            .with_cypher_endpoints(crate::CypherEndpoints::Enabled),
    );
    for (body, status) in [
        (
            json!({"query":"CREATE (:N {key:7})"}).to_string(),
            StatusCode::OK,
        ),
        (json!({"query":"RETURN 1/0"}).to_string(), StatusCode::OK),
        (
            json!({"query":"RETURN $missing"}).to_string(),
            StatusCode::BAD_REQUEST,
        ),
        ("invalid json".to_owned(), StatusCode::BAD_REQUEST),
    ] {
        let response = router
            .clone()
            .oneshot(
                Request::post("/v2/cypher/explain")
                    .body(Body::from(body))
                    .unwrap(),
            )
            .await
            .unwrap();
        assert_eq!(response.status(), status);
        let value: serde_json::Value =
            serde_json::from_slice(&to_bytes(response.into_body(), 65536).await.unwrap()).unwrap();
        if status == StatusCode::OK {
            assert!(value.get("operators").is_some());
            assert!(value.get("columns").is_none());
            assert!(value.get("rows").is_none());
        }
    }
    assert_eq!(
        db.cypher(db::cypher::Request::new("MATCH (n:N) RETURN count(*)"))
            .await
            .unwrap()
            .rows,
        vec![vec![json!(0)]]
    );
    drop(router);
    db.close().await.unwrap();
}

#[tokio::test]
async fn cypher_routing_checks_effects_before_parameter_validation() {
    let source = db::HelixDbSource::InMemory {
        database: "cypher-routing-precedence".into(),
    };
    let config = source
        .embedded_default_config()
        .with_query_telemetry(db::config::QueryTelemetry::Disabled);
    let db = Arc::new(db::HelixDB::open_with_config(source, config).await.unwrap());
    let state = state::ServerState::new(Arc::clone(&db), None, crate::StorageConfig::Memory)
        .with_cypher_endpoints(crate::CypherEndpoints::Enabled);
    let router = http::router(state.clone());
    let grpc = grpc::GrpcService::new(state);
    for (text, warm, durable, detail) in [
        ("CREATE (:N {key:$missing})", true, false, None),
        ("RETURN $missing", false, true, None),
        ("RETURN $missing", true, false, Some("MissingParameter")),
        ("RETURN missing", false, true, Some("UndefinedVariable")),
    ] {
        let body = json!({"query":text}).to_string();
        let response = router
            .clone()
            .oneshot(
                Request::post("/v2/cypher")
                    .header("x-helix-warm", warm.to_string())
                    .header("x-helix-await-durable", durable.to_string())
                    .body(Body::from(body.clone()))
                    .unwrap(),
            )
            .await
            .unwrap();
        assert_eq!(response.status(), StatusCode::BAD_REQUEST);
        let value: serde_json::Value =
            serde_json::from_slice(&to_bytes(response.into_body(), 4096).await.unwrap()).unwrap();
        let error = grpc
            .execute_cypher(tonic::Request::new(grpc::pb::QueryJsonRequest {
                body: body.into_bytes().into(),
                warm_only: warm,
                await_durable: durable,
                ..Default::default()
            }))
            .await
            .unwrap_err();
        assert_eq!(error.code(), tonic::Code::InvalidArgument);
        match detail {
            Some(detail) => {
                assert_eq!(value["details"]["detail"], detail);
                assert_eq!(value["details"]["phase"], "compile");
                let error: helix_planner::relational::QueryError =
                    serde_json::from_slice(error.details()).unwrap();
                assert_eq!(error.detail, detail);
                assert_eq!(error.phase, helix_planner::relational::ErrorPhase::Compile);
            }
            None => {
                assert_eq!(value["error"], "invalid_request_option");
                assert_eq!(
                    error
                        .metadata()
                        .get(grpc::HELIX_ERROR_CODE_METADATA)
                        .unwrap()
                        .to_str()
                        .unwrap(),
                    "invalid_request_option"
                );
            }
        }
    }
    assert_eq!(
        db.cypher(db::cypher::Request::new("MATCH (n:N) RETURN count(*)"))
            .await
            .unwrap()
            .rows,
        vec![vec![json!(0)]]
    );
    drop(router);
    drop(grpc);
    db.close().await.unwrap();
}

#[tokio::test]
async fn cypher_endpoints_are_unrouted_unless_enabled() {
    let db = Arc::new(
        db::HelixDB::open(db::HelixDbSource::InMemory {
            database: "cypher-disabled".into(),
        })
        .await
        .unwrap(),
    );
    let state = state::ServerState::new(Arc::clone(&db), None, crate::StorageConfig::Memory);
    assert_eq!(state.cypher_endpoints(), crate::CypherEndpoints::Disabled);
    let router = http::router(state.clone());
    for path in ["/v2/cypher", "/v2/cypher/explain"] {
        let response = router
            .clone()
            .oneshot(
                Request::post(path)
                    .body(Body::from(json!({"query":"RETURN 1"}).to_string()))
                    .unwrap(),
            )
            .await
            .unwrap();
        assert_eq!(response.status(), StatusCode::NOT_FOUND, "{path}");
    }
    let error = grpc::GrpcService::new(state)
        .execute_cypher(tonic::Request::new(grpc::pb::QueryJsonRequest {
            body: json!({"query":"RETURN 1"}).to_string().into_bytes().into(),
            ..Default::default()
        }))
        .await
        .unwrap_err();
    assert_eq!(error.code(), tonic::Code::Unimplemented);
    // Native queries are unaffected.
    let response = router
        .oneshot(
            Request::post("/v2/query")
                .body(Body::from(
                    serde_json::to_vec(&helix_ast::query::QueryRequest::read(
                        helix_ast::batch::read_batch()
                            .var_as("n", helix_ast::traversal::g().n_with_label("X").count())
                            .returning(["n"]),
                    ))
                    .unwrap(),
                ))
                .unwrap(),
        )
        .await
        .unwrap();
    assert_eq!(response.status(), StatusCode::OK);
    db.close().await.unwrap();
}

#[tokio::test]
async fn cypher_rejects_malformed_bodies_and_options_before_executing() {
    let db = Arc::new(
        db::HelixDB::open(db::HelixDbSource::InMemory {
            database: "cypher-rejections".into(),
        })
        .await
        .unwrap(),
    );
    let state = state::ServerState::new(Arc::clone(&db), None, crate::StorageConfig::Memory)
        .with_cypher_endpoints(crate::CypherEndpoints::Enabled);
    let router = http::router(state.clone());
    let malformed = router
        .clone()
        .oneshot(Request::post("/v2/cypher").body(Body::from("{")).unwrap())
        .await
        .unwrap();
    assert_eq!(malformed.status(), StatusCode::BAD_REQUEST);
    // A warm read option cannot apply to a modifying statement.
    let warm_write = router
        .clone()
        .oneshot(
            Request::post("/v2/cypher")
                .header("x-helix-warm", "true")
                .body(Body::from(
                    json!({"query":"CREATE (:Rejected)"}).to_string(),
                ))
                .unwrap(),
        )
        .await
        .unwrap();
    assert_eq!(warm_write.status(), StatusCode::BAD_REQUEST);
    let error = grpc::GrpcService::new(state)
        .execute_cypher(tonic::Request::new(grpc::pb::QueryJsonRequest {
            body: b"{".to_vec().into(),
            ..Default::default()
        }))
        .await
        .unwrap_err();
    assert_eq!(error.code(), tonic::Code::InvalidArgument);
    let count = router
        .oneshot(
            Request::post("/v2/cypher")
                .body(Body::from(
                    json!({"query":"MATCH (n:Rejected) RETURN count(n) AS n"}).to_string(),
                ))
                .unwrap(),
        )
        .await
        .unwrap();
    let value: serde_json::Value =
        serde_json::from_slice(&to_bytes(count.into_body(), 4096).await.unwrap()).unwrap();
    assert_eq!(value["rows"], json!([[0]]));
    db.close().await.unwrap();
}