# docs/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "HelixDB HTTP API",
    "version": "2.0.0",
    "summary": "Execute operation-tree and Cypher queries against HelixDB",
    "description": "The v2 operation-tree query contract is available on local servers and Helix Cloud gateways. Local servers additionally expose Cypher execution and read-only planning. Cypher gateway deployment is outside the local MVP contract.",
    "termsOfService": "https://www.helix-db.com/terms",
    "contact": {
      "name": "HelixDB",
      "url": "https://www.helix-db.com/developers"
    },
    "license": {
      "name": "Apache-2.0",
      "identifier": "Apache-2.0"
    }
  },
...
  "servers": [
    {
      "url": "http://localhost:6969",
      "description": "Default local HelixDB server"
    }
  ],
...
  "paths": {
    "/v2/query": {
      "post": {
        "operationId": "executeQuery",
        "summary": "Execute a HelixDB query",
        "description": "Executes one read or write batch. request_type must match the closed read or write variant under query. Local servers accept request bodies up to 16 MiB. Helix Cloud gateways accept request bodies up to 2 MiB.",
        "x-helix-request-body-limits": {
          "localBytes": 16777216,
          "helixCloudBytes": 2097152
        },
        "servers": [
          {
            "url": "http://localhost:6969",
            "description": "Default local HelixDB server"
          },
          {
            "url": "https://{gatewayHost}",
            "description": "Helix Cloud query gateway from the database connection details",
            "variables": {
              "gatewayHost": {
                "default": "cluster.helix-db.com",
                "description": "Placeholder only. Replace it with the managed gateway hostname shown in the Helix Cloud dashboard."
              }
            }
          }
        ],
        "tags": [
          "Queries"
        ],
        "security": [
          {},
          {
            "bearerAuth": []
          }
        ],
        "parameters": [
          {
            "$ref": "#/components/parameters/DatabaseId"
          },
          {
            "$ref": "#/components/parameters/WarmOnly"
          },
          {
            "$ref": "#/components/parameters/RequireWriter"
          },
          {
            "$ref": "#/components/parameters/AwaitDurable"
          }
        ],
        "requestBody": {
          "required": true,
          "description": "Maximum encoded body size: 16 MiB on a local HelixDB server and 2 MiB on a Helix Cloud gateway.",
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/QueryRequest"
              },
              "examples": {
                "read": {
                  "summary": "Count User nodes",
                  "value": {
                    "request_type": "read",
                    "query_name": "node_count",
                    "query": {
                      "read": {
                        "entries": [
                          {
                            "query": {
                              "name": "node_count",
                              "root": {
                                "count": {
                                  "input": {
                                    "nodes_where": {
                                      "predicate": {
                                        "eq": {
                                          "left": {
                                            "property": "$label"
                                          },
                                          "right": {
                                            "constant": {
                                              "string": "User"
                                            }
                                          }
                                        }
                                      }
                                    }
                                  }
                                }
                              }
                            }
                          }
                        ],
                        "returns": [
                          "node_count"
                        ]
                      }
                    }
                  }
                },
                "write": {
                  "summary": "Create one User node",
                  "value": {
                    "request_type": "write",
                    "query_name": "create_user",
                    "query": {
                      "write": {
                        "entries": [
                          {
                            "query": {
                              "name": "user",
                              "root": {
                                "add_n": {
                                  "label": "User",
                                  "properties": [
                                    [
                                      "name",
                                      {
                                        "value": {
                                          "string": "Alice"
                                        }
                                      }
                                    ]
                                  ]
                                }
                              }
                            }
                          }
                        ],
                        "returns": [
                          "user"
                        ]
                      }
                    }
                  }
                }
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Query executed successfully.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/QueryResponse"
                }
              }
            }
          },
          "204": {
            "description": "Helix Cloud completed a cache-warming read and intentionally omitted the response body."
          },
          "400": {
            "$ref": "#/components/responses/BadRequest"
          },
          "401": {
            "$ref": "#/components/responses/Unauthorized"
          },
          "402": {
            "$ref": "#/components/responses/PaymentRequired"
          },
          "403": {
            "$ref": "#/components/responses/Forbidden"
          },
          "408": {
            "$ref": "#/components/responses/RequestTimeout"
          },
          "409": {
            "$ref": "#/components/responses/Conflict"
          },
          "413": {
            "$ref": "#/components/responses/PayloadTooLarge"
          },
          "429": {
            "$ref": "#/components/responses/RateLimited"
          },
          "500": {
            "$ref": "#/components/responses/InternalError"
          },
          "503": {
            "$ref": "#/components/responses/Unavailable"
          }
        }
      }
    },
...
    "/v2/cypher": {
      "post": {
        "operationId": "executeCypher",
        "summary": "Execute a Cypher statement",
        "description": "Executes the supported Cypher read/CRUD profile. The server derives read/write effects from the statement. Modifying statements execute atomically in one write transaction. X-Helix-Warm is valid only for reads; X-Helix-Await-Durable:true is valid only for writes. This operation is available on the standalone local server started with HELIX_ENABLE_CYPHER=true; otherwise the path returns 404. Cloud gateway deployment is outside the MVP. The body limit is 16 MiB, the default deadline is 30 seconds, and the response limit is 16 MiB. Resource exhaustion returns an error instead of truncation.",
        "tags": [
          "Cypher"
        ],
        "security": [
          {}
        ],
        "x-helix-request-body-limits": {
          "localBytes": 16777216
        },
        "parameters": [
          {
            "$ref": "#/components/parameters/WarmOnly"
          },
          {
            "$ref": "#/components/parameters/RequireWriter"
          },
          {
            "$ref": "#/components/parameters/AwaitDurable"
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CypherRequest"
              },
              "examples": {
                "read": {
                  "summary": "Find people by a parameter",
                  "value": {
                    "query": "MATCH (n:Person) WHERE n.age >= $age RETURN n.name AS name",
                    "parameters": {
                      "age": 30
                    },
                    "query_name": "adults"
                  }
                },
                "write": {
                  "summary": "Create a labeled node",
                  "value": {
                    "query": "CREATE (n:Person {name:$name}) RETURN n",
                    "parameters": {
                      "name": "Ada"
                    }
                  }
                }
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Statement executed successfully, including writes without RETURN.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/CypherResponse"
                }
              }
            }
          },
          "400": {
            "description": "Malformed or oversized body, invalid request options, or a classified Cypher compile/runtime error. The local 16 MiB body limit returns invalid_request_body with this status.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/CypherError"
                }
              }
            }
          },
          "409": {
            "description": "Native storage transaction conflict.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/CypherError"
                }
              }
            }
          },
          "429": {
            "description": "Cypher resource limit exceeded; inspect details.detail.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/CypherError"
                }
              }
            }
          },
          "500": {
            "description": "Internal planner, storage, execution-control, or serialization failure.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/CypherError"
                }
              }
            }
          },
          "503": {
            "description": "Writer required, server access mode unavailable, or uncertain write commit. An uncertain outcome carries retryable:false.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/CypherError"
                }
              }
            }
          }
        }
      }
    },
    "/v2/cypher/explain": {
      "post": {
        "operationId": "explainCypher",
        "summary": "Explain a Cypher statement",
        "description": "Parses, binds, and plans one statement against the current catalog without executing it. Uses read request options even for a modifying statement. X-Helix-Await-Durable:true is invalid. This operation is available on the standalone local server started with HELIX_ENABLE_CYPHER=true; otherwise the path returns 404. Cloud gateway deployment is outside the MVP. The body limit is 16 MiB, the default deadline is 30 seconds, and the response limit is 16 MiB. Resource exhaustion returns an error instead of truncation.",
        "tags": [
          "Cypher"
        ],
        "security": [
          {}
        ],
        "x-helix-request-body-limits": {
          "localBytes": 16777216
        },
        "parameters": [
          {
            "$ref": "#/components/parameters/WarmOnly"
          },
          {
            "$ref": "#/components/parameters/RequireWriter"
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CypherRequest"
              },
              "examples": {
                "read": {
                  "summary": "Find people by a parameter",
                  "value": {
                    "query": "MATCH (n:Person) WHERE n.age >= $age RETURN n.name AS name",
                    "parameters": {
                      "age": 30
                    },
                    "query_name": "adults"
                  }
                },
                "write": {
                  "summary": "Create a labeled node (plan only)",
                  "value": {
                    "query": "CREATE (n:Person {name:$name}) RETURN n",
                    "parameters": {
                      "name": "Ada"
                    }
                  }
                }
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Statement planned without execution.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/CypherExplanation"
                }
              }
            }
          },
          "400": {
            "description": "Malformed or oversized body, invalid request options, or a classified Cypher compile/runtime error. The local 16 MiB body limit returns invalid_request_body with this status.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/CypherError"
                }
              }
            }
          },
          "429": {
            "description": "Cypher resource limit exceeded; inspect details.detail.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/CypherError"
                }
              }
            }
          },
          "500": {
            "description": "Internal planner, storage, execution-control, or serialization failure.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/CypherError"
                }
              }
            }
          },
          "503": {
            "description": "A request option requires a writer but the server is read-only.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/CypherError"
                }
              }
            }
          }
        }
      }
    }
  },
  "components": {
    "securitySchemes": {
      "bearerAuth": {
        "type": "http",
        "scheme": "bearer",
        "description": "Helix Cloud cluster API key. Local servers do not require this credential by default."
      }
    },
    "parameters": {
      "DatabaseId": {
        "name": "X-Helix-Database-Id",
        "in": "header",
        "required": false,
        "description": "Database identifier shown in Helix Cloud connection details. Required by the GA shared gateway, not needed by a standalone local server, and not allowed by a database-specific cluster-mode gateway. The legacy X-Helix-Tenant-Id alias is also accepted in GA mode.",
        "schema": {
          "type": "string",
          "minLength": 1
        }
      },
      "WarmOnly": {
        "name": "X-Helix-Warm",
        "in": "header",
        "required": false,
        "description": "Warm read execution state. Valid only for read requests.",
        "schema": {
          "type": "boolean",
          "default": false
        }
      },
      "RequireWriter": {
        "name": "X-Helix-Require-Writer",
        "in": "header",
        "required": false,
        "description": "Reject the request unless it reaches a writer-capable server.",
        "schema": {
          "type": "boolean",
          "default": false
        }
      },
      "AwaitDurable": {
        "name": "X-Helix-Await-Durable",
        "in": "header",
        "required": false,
        "description": "Flush the writer before acknowledging success. Valid only for write requests.",
        "schema": {
          "type": "boolean",
          "default": false
        }
      }
    },
    "responses": {
      "BadRequest": {
        "description": "Invalid JSON, query plan, vector input, request option, or Helix Cloud authentication header.",
        "content": {
          "application/json": {
            "schema": {
              "$ref": "#/components/schemas/QueryError"
            }
          }
        }
      },
      "Unauthorized": {
        "description": "The Helix Cloud API key is invalid.",
        "content": {
          "application/json": {
            "schema": {
              "$ref": "#/components/schemas/QueryError"
            }
          }
        }
      },
      "PaymentRequired": {
        "description": "Helix Cloud query processing is disabled because the tenant has exhausted its credit.",
        "content": {
          "application/json": {
            "schema": {
              "$ref": "#/components/schemas/QueryError"
            }
          }
        }
      },
      "Forbidden": {
        "description": "The Helix Cloud API key does not have permission to execute the query.",
        "content": {
          "application/json": {
            "schema": {
              "$ref": "#/components/schemas/QueryError"
            }
          }
        }
      },
      "RequestTimeout": {
        "description": "The Helix Cloud query exceeded its wall-clock limit.",
        "content": {
          "application/json": {
            "schema": {
              "$ref": "#/components/schemas/QueryError"
            }
          }
        }
      },
      "Conflict": {
        "description": "The transaction conflicted with concurrent work and can be retried.",
        "content": {
          "application/json": {
            "schema": {
              "$ref": "#/components/schemas/QueryError"
            }
          }
        }
      },
      "PayloadTooLarge": {
        "description": "The Helix Cloud gateway rejected a request body larger than 2 MiB before query parsing.",
        "content": {
          "application/json": {
            "schema": {
              "$ref": "#/components/schemas/QueryError"
            },
            "example": {
              "error": "payload_too_large",
              "msg": "request body exceeds the maximum allowed size"
            }
          }
        }
      },
      "RateLimited": {
        "description": "The Helix Cloud gateway rate limit was exceeded, or the database rejected a request with `index_backpressure`: a write because a vector or text index already retains its maximum unpublished work, or a whole-index strong search whose answer lies behind more than 800 unpublished changes (eventual and prefiltered searches are never rejected this way), or a strong text search that would analyze more unpublished text in its partition than its bound allows (512 MiB of analysis by default), or a strong vector search that would score more unpublished vector work than its bound allows (by default the index's retained-byte limit, which writes reach first). Eventual searches are never rejected for unpublished text or vector work. Backpressure rejects the whole request (writes before commit) and is marked `retryable`.",
        "headers": {
          "Retry-After": {
            "description": "Delay before retrying, when supplied by the gateway.",
            "schema": {
              "type": "string"
            }
          }
        },
        "content": {
          "application/json": {
            "schema": {
              "$ref": "#/components/schemas/QueryError"
            }
          }
        }
      },
      "InternalError": {
        "description": "An internal query, storage, or serialization failure occurred.",
        "content": {
          "application/json": {
            "schema": {
              "$ref": "#/components/schemas/QueryError"
            }
          }
        }
      },
      "Unavailable": {
        "description": "A required writer or ready database is unavailable.",
        "content": {
          "application/json": {
            "schema": {
              "$ref": "#/components/schemas/QueryError"
            }
          }
        }
      }
    },
    "schemas": {
      "QueryRequest": {
        "oneOf": [
          {
            "$ref": "#/components/schemas/ReadQueryRequest"
          },
          {
            "$ref": "#/components/schemas/WriteQueryRequest"
          }
        ],
        "discriminator": {
          "propertyName": "request_type",
          "mapping": {
            "read": "#/components/schemas/ReadQueryRequest",
            "write": "#/components/schemas/WriteQueryRequest"
          }
        }
      },
...
      "QueryResponse": {
        "type": "object",
        "description": "Returned variables keyed by the names listed in the batch returns array.",
        "additionalProperties": true,
        "example": {
          "node_count": [
            42
          ]
        }
      },
      "QueryError": {
        "type": "object",
        "description": "Stable error envelope returned by local servers and Helix Cloud gateways.",
        "required": [
          "error",
          "msg"
        ],
        "properties": {
          "error": {
            "type": "string",
            "description": "Stable snake_case error code."
          },
          "msg": {
            "type": "string",
            "description": "Human-readable diagnostic message."
          }
        },
        "additionalProperties": false,
        "example": {
          "error": "invalid_query_json",
          "msg": "invalid query JSON: expected value"
        }
      },
...
      "CypherRequest": {
        "type": "object",
        "additionalProperties": false,
        "required": [
          "query"
        ],
        "properties": {
          "query": {
            "type": "string",
            "description": "One statement in the supported Cypher profile; empty or unsupported statements return a classified query error."
          },
          "parameters": {
            "type": "object",
            "additionalProperties": {
              "$ref": "#/components/schemas/CypherParameterValue"
            },
            "default": {}
          },
          "query_name": {
            "type": [
              "string",
              "null"
            ],
            "description": "Optional caller-supplied query name."
          }
        }
      },
...
    }
  }
}